Add map movement and very basic path movement
The map movement is a bit sudden when moving from end-to-end, so it
might be worth trying to smooth it out a bit. The movement is very
blocky and precise. You must be directly in the middle of a block to
move in an adjacent path. The original game seems to round corners to
prevent blocky movement so that will be a needed improvement.