Fixed the bombs moving with the map
The bombs now stay in their respective location. I haven't set the grids
of the bombs yet. Also had an idea about fixing the movement a bit more.
Was thinking we could use the arrows keys just to change direction and
in the buffer function make the actual movements. We could add a
standing position that is inacted when all arrows keys are released.
Then we can add the timer to slow down the movement so he doesnt move 60
times a second. Seems like it would make everything move a bit more
smoothly.