Made movement more fluid, implemented dying.
Bomberman and baroms can die due to being in explosion, or bomberman
being in the same grid as a barom. Can now place bombs and move at the
same time. Can round corners so you don't have to be exactly in the
middle. The movement is still a bit rigid and most likely has some bugs
that will need fixing.