Updated collision detection, game over screen
After 3 deaths a screen shows the final score of the player. Collision
detection is now based on if the 32x32 box of the object overlaps
another object (i.e. bomberman and explosion). Small bug with explosions
and baroms. If the barom is in a grid that is diagonal to an explosion
and is not centered properly, it detects as overlap. To fix this we
could make the baroms be centered in the square and move only square to
square